Added a script to show number of occurences of events
The script will find all the unique events from the event-logs and will
display them along with the number of times that they occurred while
the script ran.

You can choose to execute a single benchmark, all of the benchmarks for
a given execution, or you can pass the script a file that contains a
list of the benchmarks you are interested in.